+++ src/app/(dashboard)/dashboard/[id]/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { getDashboardId } from '@/app/api/getDashboardId';
+import { getDashboardId } from '@/app/api/dashboards';
 import BarButton from '@/components/commons/button/BarButton';
 import DashBoardHeader from '@/components/commons/layout/DashboardHeader';
 import Sidemenu from '@/components/commons/layout/Sidemenu';
