+++ src/app/api/columns.ts
@@ -1,5 +1,6 @@
 import { getCookie } from '@/utils/cookies';
 import {
+  Column,
   ColumnsListResponse,
   CreateColumnRequest,
   CreateColumnResponse,
@@ -10,7 +11,45 @@ import {
 
 import { API_URL } from './baseUrl';
 
-export async function getColumns(
+// 컬럼 목록 조회
+type GetColumnsParams = {
+  dashboardId: number;
+};
+
+export type ColumnsResponse = {
+  result: string;
+  data: Column[];
+};
+
+export async function getColumns({
+  dashboardId,
+}: GetColumnsParams): Promise<Column[]> {
+  try {
+    const token = getCookie('accessToken');
+    const response = await fetch(
+      `${API_URL}/columns?dashboardId=${dashboardId}`,
+      {
+        headers: { Authorization: `Bearer ${token}` },
+      },
+    );
+
+    if (!response.ok) {
+      throw new Error(`HTTP error: ${response.status}`);
+    }
+
+    const body: ColumnsResponse = await response.json();
+    if (body.result !== 'SUCCESS') {
+      throw new Error(`API error: ${body.result}`);
+    }
+
+    return body.data;
+  } catch (err) {
+    throw new Error('데이터를 받는 중에 오류가 발생했습니다.');
+  }
+}
+
+// FIXME: 중복
+export async function getColumnList(
   dashboardId: number,
 ): Promise<ColumnsListResponse | ErrorMessage> {
   const token = getCookie('accessToken');
