+++ src/components/dashboard/modals/EditColumnModal.tsx
@@ -1,4 +1,4 @@
-import { deleteColumn, editColumn, getColumns } from '@/app/api/columns';
+import { deleteColumn, editColumn, getColumnList } from '@/app/api/columns';
 import Button from '@/components/commons/button';
 import Input from '@/components/commons/input';
 import Modal from '@/components/commons/modal';
@@ -73,7 +73,7 @@ export default function EditColumnModal({
     if (!isOpen) return;
 
     (async () => {
-      const columnRes = await getColumns(dashboardId);
+      const columnRes = await getColumnList(dashboardId);
 
       if ('message' in columnRes) {
         toast.error(columnRes.message);
